+++ src/lightwin/util/pass_beauty.py
@@ -115,9 +115,6 @@ def insert_field_map_pass_beauty_instructions(
         which_phase="phi_0_rel",
     )
     logging.info("Overwriting a ListOfElements by its beauty counterpart.")
-    logging.warning(
-        "Expected bug: all cavities will be shown as green in plots."
-    )
     accelerator.elts = elts
     return
 
@@ -163,9 +160,7 @@ def insert_transverse_matching_instructions(
         return
 
     if len(fault_scenario) > 1:
-        raise NotImplementedError(
-            "Not sure how multiple faults would interact."
-        )
+        logging.warning("Not sure how multiple faults will interact.")
     assert _is_adapted_to_pass_beauty(beam_calculator)
     assert isinstance(fault_scenario, FaultScenario)
 
@@ -183,9 +178,6 @@ def insert_transverse_matching_instructions(
         append_stem="qp_retuning",
     )
     logging.info("Overwriting a ListOfElements by its beauty counterpart.")
-    logging.warning(
-        "Expected bug: all cavities will be shown as green in plots."
-    )
     accelerator.elts = elts
     return
 
@@ -382,9 +374,7 @@ def _field_map_pass_beauty_instructions(
 
     """
     if len(fault_scenario) > 1:
-        raise NotImplementedError(
-            "Not sure how multiple faults would interact."
-        )
+        logging.warning("Not sure how multiple faults will interact.")
     fault = fault_scenario[0]
     fix_elts = fault_scenario.fix_acc.elts
     compensating = fault.compensating_elements
@@ -424,17 +414,19 @@ def _spiral2_transverse_matching_instructions(
     altered_lattices_idx = set(sorted([elt.idx["lattice"] for elt in altered]))
     altered_lattices = [fix_elts.by_lattice[i] for i in altered_lattices_idx]
 
-    steerers_diags, steerers_adjusts = (), ()
+    first_altered_lattice_idx = min(altered_lattices_idx)
+    diag_pos_for_steerers, adjusts_steerer = (), ()
     if retune_steerers:
-        steerers_quadrupoles = _get_steerers_and_qps(
-            fix_elts, altered_lattices_idx
+        steerers_quadrupoles = _map_qps_to_steerers(
+            fix_elts, first_altered_lattice_idx
         )
-        steerers_adjusts = _steerer_adjust_commands(
+
+        adjusts_steerer = _create_adjust_steerer_commands(
             steerers_quadrupoles, number=number + 100
         )
-        steerers_diags = _steerer_diag_commands(
+        diag_pos_for_steerers = _create_bpms(
             by_lattice=fix_elts.by_lattice,
-            adjust_steerers=steerers_adjusts,
+            adjust_steerers=adjusts_steerer,
             number=number + 100,
         )
 
@@ -453,7 +445,12 @@ def _spiral2_transverse_matching_instructions(
             return []
 
     instructions = sorted(
-        [*steerers_adjusts, *steerers_diags, *qp_diagnostics, *qp_adjusts],
+        [
+            *adjusts_steerer,
+            *diag_pos_for_steerers,
+            *qp_diagnostics,
+            *qp_adjusts,
+        ],
         key=lambda x: x.idx["dat_idx"],
     )
     return instructions
@@ -486,20 +483,16 @@ def _spiral2_quadrupoles(lattices: list[list[Element]]) -> list[Quad]:
     return quadrupoles
 
 
-def _get_steerers_and_qps(
-    elts: ListOfElements, altered_lattices_idx: Collection[int]
+def _map_qps_to_steerers(
+    elts: ListOfElements, first_lattice_idx: int
 ) -> dict[Steerer, Quad]:
-    """Get all steerers after first alteration, as well as associated qps."""
+    """Map quadrupoles to their steerer from ``first_lattice_idx`` and
+    onwards."""
     steerers_quadrupoles: dict[Steerer, Quad] = {}
-    first_altered = min(altered_lattices_idx)
-    # We remove two lattices:
-    # - exit of linac
-    # - last "real" lattice because the DIAG must be one lattice after
-    lattices_after_first_alteration = elts.by_lattice[first_altered:-2]
+    lattices_after_first_alteration = elts.by_lattice[first_lattice_idx:]
 
     for lattice in lattices_after_first_alteration:
         quadrupoles = filter_elts(lattice, Quad)
-
         for qp in quadrupoles:
             steerers = filter_elts(qp.influencing_instructions, Steerer)
             for steerer in steerers:
@@ -507,7 +500,7 @@ def _get_steerers_and_qps(
     return steerers_quadrupoles
 
 
-def _steerer_adjust_commands(
+def _create_adjust_steerer_commands(
     steerers_quadrupoles: Mapping[Steerer, Quad], number: int
 ) -> list[AdjustSteerer]:
     """Create adjust steerer commands."""
@@ -530,7 +523,7 @@ def _steerer_adjust_commands(
     return adjust_steerers
 
 
-def _steerer_diag_commands(
+def _create_bpms(
     by_lattice: list[list[Element]],
     adjust_steerers: list[AdjustSteerer],
     number: int,
@@ -542,13 +535,13 @@ def _steerer_diag_commands(
     for adjust in adjust_steerers:
         if isinstance(adjust, AdjustSteererBy):
             continue
+
         next_lattice_idx = adjust.number - number + 1
-        if next_lattice_idx + 1 >= len(by_lattice):
+        if next_lattice_idx >= len(by_lattice):
             continue
+
         next_lattice = by_lattice[next_lattice_idx]
         qps = filter_elts(next_lattice, Quad)
-        if len(qps) != 3:
-            continue
 
         if next_lattice_idx <= 12:
             name = "LINA-BPM"